requirements.txt
Django==4.2
asgiref==3.8.1
defusedxml==0.7.1
diff-match-patch==20230430
django-import-export==4.0.6
django-js-asset==2.2.0
django-robots==6.1
django-tinymce==4.0.0
et-xmlfile==1.1.0
gunicorn==22.0.0
mysqlclient==2.2.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.5.0
tablib==3.5.0
tabulate==0.9.0
tzdata==2024.1
django-robots
humanize
gunicorn
psycopg2-binary